# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/dashboard/hotel_statistic": {
            "get": {
                "description": "Show amount of hotelsStatistic",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Show hotelsStatistic",
                "operationId": "hotelsStatistic",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.hotelStatisticResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        },
        "/dashboard/most_loyal_hotels": {
            "get": {
                "description": "Show all dashboards mostLoyalHotels",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "dashboard"
                ],
                "summary": "Show mostLoyalHotels",
                "operationId": "mostLoyalHotels",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1.loyalHotelsResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1.response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.HotelStatistic": {
            "type": "object",
            "properties": {
                "published": {
                    "type": "integer"
                },
                "registered": {
                    "type": "integer"
                }
            }
        },
        "entity.MostLoyalHotels": {
            "type": "object",
            "properties": {
                "externalID": {
                    "type": "string"
                },
                "hotelName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "rates": {
                    "type": "integer"
                },
                "roomTypes": {
                    "type": "integer"
                },
                "totalAmount": {
                    "type": "integer"
                }
            }
        },
        "v1.hotelStatisticResponse": {
            "type": "object",
            "properties": {
                "HotelStatistic": {
                    "$ref": "#/definitions/entity.HotelStatistic"
                }
            }
        },
        "v1.loyalHotelsResponse": {
            "type": "object",
            "properties": {
                "mostLoyalHotels": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.MostLoyalHotels"
                    }
                }
            }
        },
        "v1.response": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        }
    }
}